+++ flashinfer/__init__.py
@@ -86,7 +86,7 @@ from .fused_moe import (
 from .gemm import SegmentGEMMWrapper as SegmentGEMMWrapper
 from .gemm import bmm_fp8 as bmm_fp8
 from .gemm import mm_fp4 as mm_fp4
-from .gemm import gemm_fp8 as gemm_fp8
+from .trtllm_flavored_gemm import trtllm_flavored_gemm as trtllm_flavored_gemm
 from .gemm import tgv_gemm_sm100 as tgv_gemm_sm100
 from .mla import BatchMLAPagedAttentionWrapper as BatchMLAPagedAttentionWrapper
 from .norm import fused_add_rmsnorm as fused_add_rmsnorm
