+++ flashinfer/gemm.py
@@ -2205,170 +2205,6 @@ def mm_fp4(
         )
     return out
 
-@functools.cache
-def get_trtllm_fp8_gemm_module():
-    mod = gen_trtllm_gen_gemm_module()
-    op = mod.build_and_load()
-    setup_cubin_loader(mod.get_library_path())
-
-    class TrtllmFp8GemmRunner(TunableRunner):
-        def get_valid_tactics(
-            self,
-            inputs: List[torch.Tensor],
-            profile: OptimizationProfile,
-        ) -> List[int]:
-            a_tensor_index = 0
-            b_tensor_index = 1
-
-            # NOTE : expects  A=MxK, B=NxK, out=MxN
-            a = profile.get_opt_shapes()[a_tensor_index]
-            b = profile.get_opt_shapes()[b_tensor_index]
-            m = a[0]
-            n = b[0]
-            k = a[1]
-            (
-                a,
-                b,
-                global_scale,
-                out,
-                workspace_buffer,
-            ) = inputs
-            type_e4m3 = 1
-            type_bf16 = 2
-            valid_tactics = list(
-                op.trtllm_gemm_tactics(m, n, k, type_e4m3, type_bf16, False)
-            )
-            print("valid_tactics", valid_tactics)
-            return valid_tactics
-
-        def forward(
-            self,
-            inputs: List[torch.Tensor],
-            tactic: int = -1,
-            do_preparation: bool = False,
-            **kwargs,
-        ) -> torch.Tensor:
-            (
-                a,
-                b,
-                global_scale,
-                out,
-                workspace_buffer,
-            ) = inputs
-            op.trtllm_gemm.default(
-                workspace_buffer,
-                a,
-                b,
-                None,
-                None,
-                global_scale,
-                out,
-                False,
-                False,
-                tactic,
-            )
-            return out
-
-    def gemm_runner():
-        return TrtllmFp8GemmRunner()
-
-    # Register the module
-    return SimpleNamespace(
-        gemm_runner=gemm_runner,
-    )
-
-def fp8_gemm_sm100(
-    A: torch.Tensor,
-    B: torch.Tensor,
-    global_scale: torch.Tensor,
-    out: torch.Tensor,
-    workspace_buffer: torch.Tensor,
-    backends: List[str],
-) -> None:
-    tuner = AutoTuner.get()
-    a_tensor_index = 0
-    out_tensor_index = 3
-    tuning_config = TuningConfig(
-        dynamic_tensor_specs=(
-            DynamicTensorSpec(
-                (a_tensor_index,),
-                (-2,),
-                get_last_power_of_2_num_tokens_buckets,
-                last_positive_power_of_2,
-            ),
-        ),
-        constraint_specs=(
-            ConstraintSpec(
-                out_tensor_index, -2, lambda shapes: shapes[a_tensor_index][-2]
-            ),
-        ),
-    )
-    inputs = [A, B, global_scale, out, workspace_buffer]
-    runners: List[TunableRunner] = []
-    if "trtllm" in backends:
-        runners.append(get_trtllm_fp8_gemm_module().gemm_runner())
-    runner, tactic = tuner.choose_one(
-        "fp8_gemm",
-        runners,
-        tuning_config,
-        inputs,
-    )
-
-    runner(inputs=inputs, tactic=tactic)
-
-def gemm_fp8(
-    A: torch.Tensor,
-    B: torch.Tensor,
-    global_scale: torch.Tensor,
-    out: torch.Tensor,
-    backend: Literal["trtllm", "auto"] = "trtllm",
-) -> torch.Tensor:
-    r"""GEMM FP8 optimized for low M dimension. The optimizations require special pre-processing for the input tensor B.
-    For this reason, this implementation is only recommended for scenarios where B can be pre-processed offline, once. Luckily, this
-    is often the case in inference, where B is a weight matrix that can be massaged on-load.
-    - First B has to be shuffled.
-    - Then, its layout has to be adjusted as blockwise row-major.
-      The blocksize is 128 for e4m3, and therefore, the K dimension must be divisible by 128.
-    Utilities to perform these operations are implemented in Flashinfer, and their usage is demonstrated in the tests for this operation.
-
-    Parameters
-    ----------
-    A: torch.Tensor
-        Input tensor, shape (m, k), fp8 e4m3.
-
-    B: torch.Tensor
-        Mat2 tensor, shape (n, k), fp8 e4m3.
-
-    global_scale: torch.Tensor
-        Scale tensor for the output, float.
-
-    out: Optional[torch.Tensor]
-        Out tensor, shape (m, n), bf16.
-
-    backend: Literal["trtllm", "auto"]
-        The backend to use for the operation. Defaults to ``"trtllm"``, as it is the only supported backend currently.
-
-    Returns
-    -------
-    out: torch.Tensor
-        Out tensor, shape (m, n), bf16, will be a reference to the output parameter, as only inplace is supported.
-
-    See tests/test_fp8_gemm.py for usage examples.
-    """
-    if backend == "trtllm":
-        backends = ["trtllm"]
-    elif backend == "auto":
-        backends = ["trtllm"]
-    else:
-        raise ValueError(f"Unsupported backend: {backend}")
-
-    workspace_buffer = _get_cache_buf(
-        "gemm_fp8_workspace", DEFAULT_WORKSPACE_SIZE, A.device
-    )
-
-    fp8_gemm_sm100(A, B, global_scale, out, workspace_buffer, backends)
-    return out
-
 
 def bmm_fp8(
     A: torch.Tensor,
